context: read configuration from and write logs and temp files to private app storage
The external app public storage directory is no longer reliably accessible on Android 11 and above.
This makes editing the configuration and accessing the log files impossible in some cases.

Let's move to the app private storage, to be made accessible to the user by some other mean.

This has the benefit of also protecting the private keys that need to be stored encrypted otherwise.
We also split the configuration and cache directory into specialised sub-directories.

GitHub: related to #103