app: add configuration directory migrator
This adds an automatic configuration migrator which moves the configuration files previously located
on the external public-ish directory to the private app storage, so that existing users can seamlessly
transition to the next version.

The cache directory containing the logs is cleared instead of being migrated, because those files are
considered disposable.

GitHub: related to #103